Add TheOddsApi SDK and remove old App files
This commit removes the original App and AppTest files and introduces a new SDK for interacting with The Odds API. This includes classes for network requests, URL utilities, and data structures for sports data, as well as a new logger configuration file. It also updates the project groupId in the pom.xml and adds new dependencies.