Feature/11 release version 100 (#12)
* Update method references in test comments

The method references in test case comments for 'getSports' and 'getOdds' in TheOddsApiTest were updated. The updated reference for 'getSports' includes 'Boolean' type and 'getOdds' includes the full package name for 'Sport'.

* Update version in pom.xml

Changed the version from 1.0-SNAPSHOT to 1.0.0 in the pom.xml file. This signifies the transition from a development version to a release version.